Enhance logic to find all missing messages
We cannot stop looking for a message as soon as we reach the exact number. If we are looking for 1 missing message for a specific queue and there are 100 in a single file for that queue. The message we are looking for will be the last one in the fine, not the first one. So it is preferable to get all related message in the file instead.

The result is that we grab more messages that required, but this is not a problem.

Also, we need to make a distinction between messages coming from the push API vs the crawlers. So I added something at the beginning  of the file to make that distinction.